One more check

+++ wrkzcoin_tipbot/store.py
@@ -133,10 +133,11 @@ async def sql_register_user(userID, coin: str = None):
                 cur.execute(sql, (str(userID), COIN_NAME))
                 result = cur.fetchone()
                 # recreate new address
-                if 'paymentid' not in result or result['paymentid'] is None or (len(result['paymentid']) != 16 and len(result['paymentid']) != 64):
-                    sql = """ DELETE FROM """+coin.lower()+"""_user_paymentid WHERE `user_id`=%s AND `coin_name` = %s LIMIT 1 """
-                    cur.execute(sql, (str(userID), COIN_NAME))
-                    result = None
+                if result is not None:
+                    if 'paymentid' not in result or result['paymentid'] is None or (len(result['paymentid']) != 16 and len(result['paymentid']) != 64):
+                        sql = """ DELETE FROM """+coin.lower()+"""_user_paymentid WHERE `user_id`=%s AND `coin_name` = %s LIMIT 1 """
+                        cur.execute(sql, (str(userID), COIN_NAME))
+                        result = None
             elif coin in ENABLE_COIN_DOGE:
                 sql = """ SELECT user_id, balance_wallet_address, user_wallet_address FROM """+coin.lower()+"""_user
                           WHERE `user_id`=%s LIMIT 1 """
